Preditor: Move Flowgraph Editor to its own library
