+++ Preditor/CMakeLists.txt
@@ -8,6 +8,7 @@ add_subdirectory(Common)
 add_subdirectory(Assets)
 add_subdirectory(EditTools)
 add_subdirectory(Engine)
+add_subdirectory(FGE)
 add_subdirectory(GameEditor)
 add_subdirectory(Input)
 add_subdirectory(LevelEditor)
