+++ Preditor/FGE/Edge.h
